# TrackedArray: implement `mean`