@iblis17 2017-10-30
Oh, I missed this:
```patch
diff --git a/src/tracker/lib.jl b/src/tracker/lib.jl
index 9f3adc6..83be737 100644
--- a/src/tracker/lib.jl
+++ b/src/tracker/lib.jl
@@ -57,8 +57,12 @@ back(::typeof(sum), Δ, xs::TrackedArray, dim...) = back(xs, similar(xs.data) .=
 Base.maximum(xs::TrackedArray, args...) = maximum(xs.data, args...)
 Base.findfirst(xs::TrackedArray, args...) = findfirst(xs.data, args...)
 
+Base.mean(xs::TrackedArray) =
+  TrackedArray(Call(mean, xs), toarray(xs.data, mean(xs.data)))
 Base.mean(xs::TrackedArray, region) = TrackedArray(Call(mean, xs, region))
 
+back(::typeof(mean), Δ, xs::TrackedArray) =
+  back(xs, similar(xs.data) .= Δ ./ length(xs.data))
 back(::typeof(mean), Δ, xs::TrackedArray, region) =
   back(xs, similar(xs.data) .= Δ ./ prod(size(xs.data, region...)))
```

----

```julia
julia> x
Tracked 4-element Array{Float64,1}:
 1.0
 2.0
 3.0
 4.0

julia> @benchmark Flux.Tracker.back!(sum($x) ./ length($x))
BenchmarkTools.Trial: 
  memory estimate:  2.97 KiB
  allocs estimate:  62
  --------------
  minimum time:     19.617 μs (0.00% GC)
  median time:      20.383 μs (0.00% GC)
  mean time:        21.237 μs (2.01% GC)
  maximum time:     4.348 ms (97.99% GC)
  --------------
  samples:          10000
  evals/sample:     1

julia> @benchmark Flux.Tracker.back!(mean($x))
BenchmarkTools.Trial: 
  memory estimate:  416 bytes
  allocs estimate:  13
  --------------
  minimum time:     1.002 μs (0.00% GC)
  median time:      1.036 μs (0.00% GC)
  mean time:        1.144 μs (4.75% GC)
  maximum time:     277.461 μs (98.25% GC)
  --------------
  samples:          10000
  evals/sample:     11
```