@MikeInnes 2017-10-31
Squashed in https://github.com/FluxML/Flux.jl/commit/3d8b7250aeff4e0f600e39af6372e471df8240ac. Thanks!